Fix errors

+++ app/src/main/java/ch/epfl/sweng/spotOn/singletonReferences/StorageRef.java
@@ -1,6 +1,7 @@
 package ch.epfl.sweng.spotOn.singletonReferences;
 
 import android.support.annotation.NonNull;
+import android.util.Log;
 
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
@@ -23,7 +24,13 @@ public class StorageRef {
     }
 
     public static void deletePictureFromStorage(String pictureID){
-        mMediaDirectory.child(pictureID+".jpg").delete();
+        if(pictureID == null)
+        {
+            Log.e("Error in StorageRef","deletePictureFromStorage pictureId is null");
+        }
+        else {
+            mMediaDirectory.child(pictureID + ".jpg").delete();
+        }
     }
 
 // CONSTUCTOR FOR SINGLETON
